[ubicloud] Introduce ubicloud/cache package
The `actions/toolkit` repository provides a set of packages to make
creating actions easier. We've forked this repository and refactored the
`actions/cache` package into the `ubicloud/cache` package. This updated
version can interact with Ubicloud's cache service. We aimed to retain
as much of the original code as possible, but differences in the
architecture of Azure Blob Storage and S3 compatible storage
necessitated some changes.

This commit only modifies branding literals and the package name.
Subsequent commits will implement changes to the code.

To stay in sync with GitHub's release cycle for the cache package, I've
decided to keep our versioning as similar as possible. The only
difference is that I've multiplied the patch version by `100`. So, if
the original version is `3.2.4`, the corresponding version in the
`ubicloud/cache` package will be `3.2.401`. This approach simplifies
version tracking and allows us to release multiple build versions for
the same version of the original package.

I've also renamed `ACTIONS_CACHE_URL` to `UBICLOUD_CACHE_URL` and
`ACTIONS_RUNTIME_TOKEN` to `UBICLOUD_RUNTIME_TOKEN`. We pass these
variables to our runners during environment setup.