[ubicloud] Make saveCache Ubicloud compatible
`saveCache` consists of three main components: reserveCache, uploadFile,
and commitCache. We use presigned URLs to enable the cache client to
interact with blob storage without passing the blob storage credentials.
The design of Azure blob storage and S3 compatible API cache differs in
terms of concurrent file uploading. As Ubicloud uses an S3 compatible
API, we need to make saveCache compatible with Ubicloud.

To enable concurrent file uploading, we first send a
`CreateMultipartUpload` request to the blob storage. This request
returns an `UploadId`, which we then use to upload the file in parts.
Each uploaded part generates an `ETag`, which is essential for
completing the upload. After uploading all parts, we send a
`CompleteMultipartUpload` request, which includes all the `ETags` of the
parts and the `UploadId`. This finalizes the upload, making the file
available in the blob storage.

When the cache client sends a `reserveCache` request to the Ubicloud
cache service, it triggers a `CreateMultipartUpload` request to the blob
storage. This returns the `UploadId` and a presigned URL for each part
to the cache client. The cache client then uses these presigned URLs to
upload the file parts to blob storage and collects the `ETags` from each
response. After all parts are uploaded, the cache client sends a
`commitCache` request to the Ubicloud cache service, including the
`ETags` and `UploadId`. The cache service then sends a
`CompleteMultipartUpload` request to the blob storage to finalize the
upload.