[ubicloud] Make downloadCacheHttpClientConcurrent Ubicloud compatible
We also use presigned URLs to restore caches from blob storage. The
design of Azure blob storage and the S3 compatible API are more similar
in terms of file downloading than uploading. Since Ubicloud uses an S3
compatible API, we need to ensure `restoreCache` is compatible with
Ubicloud.

When the `restoreCache` method of the cache client called, it first
sends a `getCacheEntry` request to the Ubicloud cache service. The
response includes `archiveLocation`, a presigned URL used to download
the cache. If this location is a URL from Azure Blob Storage, the
download occurs concurrently. Otherwise, it downloads as a single
thread.

The current implementation of `downloadCacheHttpClientConcurrent` is not
compatible with the S3 compatible API. The main difference lies in how
we obtain the file size. In Azure Blob Storage, we can get the file size
by sending a HEAD request to the URL. However, with the S3 compatible
API, we need to send a GET request to the URL with a `Range: bytes=0-1`
header to get the file size. This PR ensures
`downloadCacheHttpClientConcurrent` is compatible with the S3 compatible
API.